+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 =====
 
 * Added support for default text and background color sequences (resetting color changes) - [@ejelly](https://github.com/ejelly).
+* Added a new color map, gnome-terminal - [@javawizard](https://github.com/javawizard).
 
 0.4.0
 =====